test-story
This is an example story with 6 requirements. Each requirement needs to be tested. Content of dropdowns needs to be verified. Input of age should be verified. Display and update of the available plans should be verified. Ability to display available plans should be verified too. 
Please, suggest scenarios in a form of feature file.